src/requirements.txt
# Django Backend Dependencies

# Core Framework
Django>=4.2,<5.0
djangorestframework == 3.15.2
djangorestframework-simplejwt>=5.3.0

# Environment Variables
python-dotenv>=1.0.0

# Admin Theme
django-unfold>=0.78.0

# Rich Text Editor
django-ckeditor-5>=0.2.13
Pillow>=10.0.0

# Excel Import/Export
openpyxl>=3.1.0

# Cryptography (RSA key generation)
cryptography>=41.0.0

# Storage Backend (S3/R2/OSS compatible)
django-storages[s3]>=1.14.2
boto3>=1.34.0

# Production Dependencies (uncomment when needed)
# gunicorn>=21.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter
django-redis>=5.4.0  # Redis cache backend

# Development Dependencies
# django-debug-toolbar>=4.0.0
ruff>=0.8.0  # Python linter and formatter

